+++ restApiFiles/Categoria.java
@@ -1,0 +1,29 @@
+package restApiFiles;
+
+public class Categorias {
+    //Spring@column
+    int idCategoria;
+    //Spring@column
+    String name;
+
+    public Categorias(int idCategoria, String name) {
+        this.idCategoria = idCategoria;
+        this.name = name;
+    }
+
+    public int getidCategoria() {
+        return idCategoria;
+    }
+
+    public void setidCategoria(int idCategoria) {
+        this.idCategoria = idCategoria;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
