--- restApiFiles/RestTest.java
@@ -1,8 +1,0 @@
-package restApiFiles;
-
-public class RestTest {
-
-    public RestTest() {
-
-    }
-}
